Battle Player modularization pt.1
A playable character (a.k.a. Player) should be dragged & dropped into
the list of Nodes of a certain Battle, then have its name referenced in
the Battle’s script.
This can ease the shaping of any battle, and even the easy
customization of them.

Changes:
- Reorganized Sound Effects (still needs better reorganization, but
this will do for now)
- Created an instance-able Scene out of Kiryoku. In case we need to
edit his behavior (i.e. Animations), we must do so in his own Scene