Rewriting SceneLoader pt.1
Old SceneLoader wasn’t working (and missed a lot of features). Loading
screens are too long, most likely something to do with the fact that
Godot is loading dozens of .png files as sprites (which is not a
healthy decision). If not, I have to find out what’s wrong.

For now, let’s just fix what’s broken.

Changes:
- Rewrote SceneLoader with minor differences
- Left old SceneLoader script for debugging purposes
- Temporarily made the global script ignore Debug input (for some
reason, it’s not checking for the combination of the presses but
IMMEDIATELY for the Letter Key.