Introducing Cinema pt.1
Cinema is where Cinematic cutscenes in the shape of videos shall
appear. In the future, this can be expanded to feature things like
subtitles, fade in/out effects and select few other.
To use it, one must create a new Scene and instance Cinema.tscn as the
root Node. Then, fill in the data available in the root Node.

Changes:
- Added Cinema Scene and Script
- Recreated Aqua cutscene with new Cinema
- Minor project modifications