Introducing Cinema pt.2
New Feature: Subtitles!
Will only accept special .srt files.
Meaning: Instead of the text that it should directly transmit, the .srt
files should contain the ID of the translation, that shall be parsed at
Subtitle-Parse-Time.
This ensures that the text shall be translated to whichever language
the System is in (or to the language set in the Game’s settings).

Changes:
- Added subtitles to Cinema
- Added subtitles for Aqua.tscn
- Relocated a couple translations