Input redux pt.1
Handling InputEvents under _input() determines what to do with that
specific event, so it can take as long as it wants (a.k.a. one action
can call a long function).

This doesn’t work if we want to work with as many keys at the same time
as possible, in which case we’ll have to properly separate single-press
actions with multi-key/multi-touch, even if they’re in the same
_input() call.

Changes:
- Simplified “Guard” code (by a lot)
- Action shortcut in global’s _input()
- Minor code rearrangements