Prepare for Battle! pt.2
If we are to use Godot to its full potential, we must forget manual
repositioning completely and adhere to the Physics engine.
We must also write the most adaptable code we can. Instead of using
hardcoded actions (guarding, attacking, etc…), we can now create as
many as we want.
This commit breaks a couple of things that must be fixed in the next
one.

Changes:
- Battlers should now be of type KinematicBody2D
- I take back what I said and must let the Battler be in immediate
control (and not let a higher-up control from above)
- Switched Kiryoku.tscn type to KinematicBody2D
- Discriminated sides completely. set_flip_h() is used instead
- Fixed Kiryoku's animations
- Temporarily broken Walk cycle
- Removed Attack and Guard's hard-coded actions
- Created instance-able Battle_Action.gd. A modular context that
controls its own set action and the ComboTimer