Profiling via PlayerData pt.1
When loading a game (with New Game or Load), PlayerData should be the
first loaded to /root. This then-Global-node will serve as a data bank
to keep on par with the Player's inventory, trophy, etc…
Note that, as soon as this is removed from root, all of its data is
lost!

Changes:
- Created minimal PlayerData script (currently does nothing)
- Added loading and unloading of nodes/resources at root level.
- Minor code improvements