BattlePlan: Enemy HP bars overhaul
Many great improvements, mainly less dependency on fixed PNG files.
This translates to more flexibility when scaling bars and restriction
on the amount of HP an enemy can have.

However, I’m kind of liking the idea more and more of saving the
enemy’s health on this node in order to avoid spamming the update()
method, which is getting kind of complex and may be a performance hit
in the long run.
The biggest problem this causes is that this data won’t be abstract
anymore. I’ll have to think of something.

Changes:
- Minimized (and even removed a few) Enemy HP textures
- Switched Enemy HP nodes type from TextureProgress to ProgressBar
- Instancing HP layers whenever necessary
- Restricting number of layers to 20 in order to avoid ridiculous
amounts of HP (although this can be removed for kicks and giggles)
- Due to these changes, a HP bar delimiter (in cases where a bar is <
100) was removed, but will be worked on
- Locked a few nodes