MapNPC ---> MapEvent
In the broadest sense, a map event should prioritize walkable and
pickable area over collisions, in case we’d like to add an action on an
empty space (e.g. Save points, or scripted sequences).

This makes it far more powerful while still maintaining the core
functionality, but we now have to be careful with MapEvent’s script
variables, namely their amount and the weird mix that’s going on.

If we want to add collisions, all we need to do is add a StaticBody2D
and a collision as its child.

Changes:
- MapNPC —> MapEvent
- Switched type to Area2D
- Added Touch NodePath variable to support MapPlayer passthroughs
- Added specificity of events
- Fixed MapSave with the changes above